// src/lib/api/types.ts
export type JobType = (typeof JobType)[keyof typeof JobType];

export const JobType = {
  job: "job",
  internship: "internship",
} as const;

export interface Job {
  id: string;
  title: string;
  company: string;
  companyLogo?: string;
  location: string;
  type: JobType;
  salary?: string;
  description: string;
  requirements?: string[];
  tags?: string[];
  employerId: string;
  createdAt: string;
  remote?: boolean;
}